# Final_API/api.yml
openapi: 3.0.2
info:
  title: Product Search API
  version: '1.0'
servers:
  - url: http://localhost:5000 
paths:
  /search:
    get:
      summary: Search for products
      parameters:
        - name: title
          in: query
          required: true
          schema:
            type: string
          description: The search query for products
        - name: maxPages
          in: query
          schema:
            type: integer
            minimum: 1
            default: 20
          description: The maximum number of pages to scrape (default 20)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    title:
                      type: string
                      description: The title of the product
                    price:
                      type: number
                      description: The price of the product
                    image:
                      type: string
                      description: The URL of the product image
    post:
      summary: Search for products using POST
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: The search query for products
                maxPages:
                  type: integer
                  minimum: 1
                  default: 20
                  description: The maximum number of pages to scrape (default 20)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    title:
                      type: string
                      description: The title of the product
                    price:
                      type: number
                      description: The price of the product
                    image:
                      type: string
                      description: The URL of the product image
    delete:
      summary: Clear search results
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string